Reaction-Diffusion redesign
Redesigned Reaction-Diffusion to support attributes for better performance.
