+++ weight_tools.py
@@ -43,17 +43,8 @@ from statistics import mean, stdev
 from mathutils import Vector
 from mathutils.kdtree import KDTree
 from numpy import *
-try: from .numba_functions import numba_reaction_diffusion, numba_reaction_diffusion_anisotropic, integrate_field
-except: pass
-#from .numba_functions import integrate_field
-#from .numba_functions import numba_reaction_diffusion
 try: import numexpr as ne
 except: pass
-
-# Reaction-Diffusion cache
-from pathlib import Path
-import random as rnd
-import string
 
 from bpy.types import (
         Operator,
@@ -74,244 +65,11 @@ from bpy.props import (
 
 from .utils import *
 
-def reaction_diffusion_add_handler(self, context):
-    # remove existing handlers
-    reaction_diffusion_remove_handler(self, context)
-    # add new handler
-    bpy.app.handlers.frame_change_post.append(reaction_diffusion_scene)
-
-def reaction_diffusion_bake_geometry(self, context):
-    props = context.object.reaction_diffusion_settings
-    if props.bake_geometry:
-        props.update_baked_geometry = True
-
-def reaction_diffusion_remove_handler(self, context):
-    # remove existing handlers
-    old_handlers = []
-    for h in bpy.app.handlers.frame_change_post:
-        if "reaction_diffusion" in str(h):
-            old_handlers.append(h)
-    for h in old_handlers: bpy.app.handlers.frame_change_post.remove(h)
-
 class formula_prop(PropertyGroup):
     name : StringProperty()
     formula : StringProperty()
     float_var : FloatVectorProperty(name="", description="", default=(0, 0, 0, 0, 0), size=5)
     int_var : IntVectorProperty(name="", description="", default=(0, 0, 0, 0, 0), size=5)
-
-class reaction_diffusion_prop(PropertyGroup):
-    run : BoolProperty(default=False, update = reaction_diffusion_add_handler,
-        description='Compute a new iteration on frame changes. Currently is not working during  Render Animation')
-
-    time_steps : IntProperty(
-        name="Steps", default=10, min=0, soft_max=50,
-        description="Number of Steps")
-
-    dt : FloatProperty(
-        name="dt", default=0.5, min=0, soft_max=1,
-        description="Time Step")
-
-    diff_a : FloatProperty(
-        name="Diff A", default=0.1, min=0, soft_max=2, precision=3,
-        description="Diffusion A")
-
-    diff_b : FloatProperty(
-        name="Diff B", default=0.05, min=0, soft_max=2, precision=3,
-        description="Diffusion B")
-
-    f : FloatProperty(
-        name="f", default=0.055, soft_min=0.01, soft_max=0.06, precision=4, step=0.05,
-        description="Feed Rate")
-
-    k : FloatProperty(
-        name="k", default=0.062, soft_min=0.035, soft_max=0.065, precision=4, step=0.05,
-        description="Kill Rate")
-
-    diff_mult : FloatProperty(
-        name="Scale", default=1, min=0, soft_max=1, max=10, precision=2,
-        description="Multiplier for the diffusion of both substances")
-
-    vertex_group_diff_a : StringProperty(
-        name="Diff A", default='',
-        description="Vertex Group used for A diffusion")
-
-    vertex_group_diff_b : StringProperty(
-        name="Diff B", default='',
-        description="Vertex Group used for B diffusion")
-
-    vertex_group_scale : StringProperty(
-        name="Scale", default='',
-        description="Vertex Group used for Scale value")
-
-    vertex_group_f : StringProperty(
-        name="f", default='',
-        description="Vertex Group used for Feed value (f)")
-
-    vertex_group_k : StringProperty(
-        name="k", default='',
-        description="Vertex Group used for Kill value (k)")
-
-    vertex_group_brush : StringProperty(
-        name="Brush", default='',
-        description="Vertex Group used for adding/removing B")
-
-    invert_vertex_group_diff_a : BoolProperty(default=False,
-        description='Inverte the value of the Vertex Group Diff A')
-
-    invert_vertex_group_diff_b : BoolProperty(default=False,
-        description='Inverte the value of the Vertex Group Diff B')
-
-    invert_vertex_group_scale : BoolProperty(default=False,
-        description='Inverte the value of the Vertex Group Scale')
-
-    invert_vertex_group_f : BoolProperty(default=False,
-        description='Inverte the value of the Vertex Group f')
-
-    invert_vertex_group_k : BoolProperty(default=False,
-        description='Inverte the value of the Vertex Group k')
-
-    min_diff_a : FloatProperty(
-        name="Min Diff A", default=0.1, min=0, soft_max=2, precision=3,
-        description="Min Diff A")
-
-    max_diff_a : FloatProperty(
-        name="Max Diff A", default=0.1, min=0, soft_max=2, precision=3,
-        description="Max Diff A")
-
-    min_diff_b : FloatProperty(
-        name="Min Diff B", default=0.1, min=0, soft_max=2, precision=3,
-        description="Min Diff B")
-
-    max_diff_b : FloatProperty(
-        name="Max Diff B", default=0.1, min=0, soft_max=2, precision=3,
-        description="Max Diff B")
-
-    min_scale : FloatProperty(
-        name="Scale", default=0.35, min=0, soft_max=1, max=10, precision=2,
-        description="Min Scale Value")
-
-    max_scale : FloatProperty(
-        name="Scale", default=1, min=0, soft_max=1, max=10, precision=2,
-        description="Max Scale value")
-
-    min_f : FloatProperty(
-        name="Min f", default=0.02, min=0, soft_min=0.01, soft_max=0.06, max=0.1, precision=4, step=0.05,
-        description="Min Feed Rate")
-
-    max_f : FloatProperty(
-        name="Max f", default=0.055, min=0, soft_min=0.01, soft_max=0.06, max=0.1, precision=4, step=0.05,
-        description="Max Feed Rate")
-
-    min_k : FloatProperty(
-        name="Min k", default=0.035, min=0, soft_min=0.035, soft_max=0.065, max=0.1, precision=4, step=0.05,
-        description="Min Kill Rate")
-
-    max_k : FloatProperty(
-        name="Max k", default=0.062, min=0, soft_min=0.035, soft_max=0.065, max=0.1, precision=4, step=0.05,
-        description="Max Kill Rate")
-
-    brush_mult : FloatProperty(
-        name="Mult", default=0.5, min=-1, max=1, precision=3, step=0.05,
-        description="Multiplier for brush value")
-
-    bool_mod : BoolProperty(
-        name="Use Modifiers", default=False,
-        description="Read modifiers affect the vertex groups")
-
-    bool_cache : BoolProperty(
-        name="Use Cache", default=False,
-        description="Read modifiers affect the vertex groups")
-
-    cache_frame_start : IntProperty(
-        name="Start", default=1,
-        description="Frame on which the simulation starts")
-
-    cache_frame_end : IntProperty(
-        name="End", default=250,
-        description="Frame on which the simulation ends")
-
-    cache_dir : StringProperty(
-        name="Cache directory", default="", subtype='FILE_PATH',
-        description = 'Directory that contains Reaction-Diffusion cache files'
-        )
-
-    update_weight_a : BoolProperty(
-        name="Update Vertex Group A", default=True,
-        description="Tranfer Cache to the Vertex Groups named A")
-
-    update_weight_b : BoolProperty(
-        name="Update Vertex Group B", default=True,
-        description="Tranfer Cache to the Vertex Groups named B")
-
-    update_colors_a : BoolProperty(
-        name="Update Vertex Color A", default=False,
-        description="Tranfer Cache to the Vertex Color named A")
-
-    update_colors_b : BoolProperty(
-        name="Update Vertex Color B", default=False,
-        description="Tranfer Cache to the Vertex Color named B")
-
-    update_colors : BoolProperty(
-        name="Update Vertex Color AB", default=False,
-        description="Tranfer Cache to the Vertex Color named AB")
-
-    update_uv : BoolProperty(
-        name="Update UV", default=False,
-        description="Tranfer Cache to the UV Map Layer named AB")
-
-    normalize : BoolProperty(
-        name="Normalize values", default=False,
-        description="Normalize values from 0 to 1")
-
-    fast_bake : BoolProperty(
-        name="Fast Bake", default=True,
-        description="Do not update modifiers or vertex groups while baking. Much faster!")
-
-    bake_geometry : BoolProperty(
-        name="Bake Geometry", default=False, update = reaction_diffusion_bake_geometry,
-        description="Bake geometry data. Improve the real-time speed. If vector fields or geometry change, it must be updated. Disable and enable to update the data")
-
-    update_baked_geometry : BoolProperty(
-        name="Update Baked Geometry", default=False,
-        description="Force to update geometry data on the next iteration.")
-
-    vector_field_mode : EnumProperty(
-            items=(
-                ('NONE', "None", "Isotropic Reaction-Diffusion"),
-                ('VECTOR', "Vector", "Uniform vector"),
-                ('OBJECT', "Object", "Orient the field with a target object's Z"),
-                ('GRADIENT', "Gradient", "Gradient vertex group"),
-                ('XYZ', "x, y, z", "Vector field defined by vertex groups 'x', 'y' and 'z'"),
-                ('VECTOR_ATTRIBUTE', "Vector Field", "'VectorField' attribute (Vertex > Vector)")
-                ),
-            default='NONE',
-            name="Vector Field controlling the direction of the Reaction-Diffusion",
-            update = reaction_diffusion_bake_geometry
-            )
-
-    anisotropy : FloatProperty(
-        name="Anisotropy", default=0.5, min=0, max=1, precision=2,
-        description="Influence of the Vector Field")
-
-    vector : FloatVectorProperty(
-        name='Vector', description='Constant Vector', default=(0.0, 0.0, 1.0),
-        update = reaction_diffusion_bake_geometry)
-
-    perp_vector_field : BoolProperty(default=False,
-        description='Use the perpendicular direction',
-        update = reaction_diffusion_bake_geometry)
-
-    vector_field_object : PointerProperty(
-        type=bpy.types.Object,
-        name="",
-        description="Target Object",
-        update = reaction_diffusion_bake_geometry
-        )
-
-    vertex_group_gradient : StringProperty(
-        name="Gradient", default='',
-        description="Vertex Group for the gradient vector field",
-        update = reaction_diffusion_bake_geometry)
 
 from numpy import *
 def compute_formula(ob=None, formula="rx", float_var=(0,0,0,0,0), int_var=(0,0,0,0,0)):
@@ -877,110 +635,6 @@ class weight_laplacian(Operator):
         bpy.ops.object.mode_set(mode='WEIGHT_PAINT')
         bm.free()
         return {'FINISHED'}
-
-
-class reaction_diffusion(Operator):
-    bl_idname = "object.reaction_diffusion"
-    bl_label = "Reaction Diffusion"
-    bl_description = ("Run a Reaction-Diffusion based on existing Vertex Groups: A and B")
-    bl_options = {'REGISTER', 'UNDO'}
-
-    steps : IntProperty(
-        name="Steps", default=10, min=0, soft_max=50,
-        description="Number of Steps")
-
-    dt : FloatProperty(
-        name="dt", default=0.5, min=0, soft_max=1,
-        description="Time Step")
-
-    diff_a : FloatProperty(
-        name="Diff A", default=1, min=0, soft_max=2,
-        description="Diffusion A")
-
-    diff_b : FloatProperty(
-        name="Diff B", default=0.5, min=0, soft_max=2,
-        description="Diffusion B")
-
-    f : FloatProperty(
-        name="f", default=0.055, min=0, soft_min=0.01, soft_max=0.06, max=0.1, precision=4,
-        description="Feed Rate")
-
-    k : FloatProperty(
-        name="k", default=0.062, min=0, soft_min=0.035, soft_max=0.065, max=0.1, precision=4,
-        description="Kill Rate")
-
-    bounds_string = ""
-
-    frame = None
-
-    @classmethod
-    def poll(cls, context):
-        return len(context.object.vertex_groups) > 0
-
-
-    def execute(self, context):
-        #bpy.app.handlers.frame_change_post.remove(reaction_diffusion_def)
-        reaction_diffusion_add_handler(self, context)
-        set_animatable_fix_handler(self, context)
-        try: ob = context.object
-        except:
-            self.report({'ERROR'}, "Please select an Object")
-            return {'CANCELLED'}
-
-        me = ob.data
-        bm = bmesh.new()
-        bm.from_mesh(me)
-        bm.edges.ensure_lookup_table()
-
-        # store weight values
-        a = []
-        b = []
-        for v in me.vertices:
-            try:
-                a.append(ob.vertex_groups["A"].weight(v.index))
-            except:
-                a.append(0)
-            try:
-                b.append(ob.vertex_groups["B"].weight(v.index))
-            except:
-                b.append(0)
-
-        a = array(a)
-        b = array(b)
-        f = self.f
-        k = self.k
-        diff_a = self.diff_a
-        diff_b = self.diff_b
-        dt = self.dt
-        n_verts = len(bm.verts)
-
-        for i in range(self.steps):
-
-            lap_a = zeros((n_verts))#[0]*n_verts
-            lap_b = zeros((n_verts))#[0]*n_verts
-            for e in bm.edges:
-                id0 = e.verts[0].index
-                id1 = e.verts[1].index
-                lap_a[id0] += a[id1] - a[id0]
-                lap_a[id1] += a[id0] - a[id1]
-                lap_b[id0] += b[id1] - b[id0]
-                lap_b[id1] += b[id0] - b[id1]
-            ab2 = a*b**2
-            a += (diff_a*lap_a - ab2 + f*(1-a))*dt
-            b += (diff_b*lap_b + ab2 - (k+f)*b)*dt
-
-            for i in range(n_verts):
-                ob.vertex_groups['A'].add([i], a[i], 'REPLACE')
-                ob.vertex_groups['B'].add([i], b[i], 'REPLACE')
-            ob.vertex_groups.update()
-            ob.data.update()
-
-            bpy.ops.wm.redraw_timer(type='DRAW_WIN_SWAP', iterations=1)
-
-        bpy.ops.object.mode_set(mode='WEIGHT_PAINT')
-        bm.free()
-        return {'FINISHED'}
-
 
 class edges_deformation(Operator):
     bl_idname = "object.edges_deformation"
@@ -2547,7 +2201,6 @@ class TISSUE_PT_weight(Panel):
         col.operator("object.weight_contour_mask", icon="MOD_MASK")
         col.separator()
         col.label(text="Simulations:")
-        #col.operator("object.reaction_diffusion", icon="MOD_OCEAN")
         col.operator("object.start_reaction_diffusion",
                     icon="EXPERIMENTAL",
                     text="Reaction-Diffusion")
@@ -2561,1181 +2214,6 @@ class TISSUE_PT_weight(Panel):
             text="Convert to Colors")
         col.operator("object.vertex_group_to_uv", icon="UV",
             text="Convert to UV")
-
-class start_reaction_diffusion(Operator):
-    bl_idname = "object.start_reaction_diffusion"
-    bl_label = "Start Reaction Diffusion"
-    bl_description = ("Run a Reaction-Diffusion based on existing Vertex Groups: A and B")
-    bl_options = {'REGISTER', 'UNDO'}
-
-    run : BoolProperty(
-        name="Run Reaction-Diffusion", default=True, description="Compute a new iteration on frame changes")
-
-    time_steps : IntProperty(
-        name="Steps", default=10, min=0, soft_max=50,
-        description="Number of Steps")
-
-    dt : FloatProperty(
-        name="dt", default=0.5, min=0, soft_max=1,
-        description="Time Step")
-
-    diff_a : FloatProperty(
-        name="Diff A", default=0.18, min=0, soft_max=2,
-        description="Diffusion A")
-
-    diff_b : FloatProperty(
-        name="Diff B", default=0.09, min=0, soft_max=2,
-        description="Diffusion B")
-
-    f : FloatProperty(
-        name="f", default=0.055, min=0, soft_min=0.01, soft_max=0.06, max=0.1, precision=4,
-        description="Feed Rate")
-
-    k : FloatProperty(
-        name="k", default=0.062, min=0, soft_min=0.035, soft_max=0.065, max=0.1, precision=4,
-        description="Kill Rate")
-
-    @classmethod
-    def poll(cls, context):
-        return context.object.type == 'MESH' and context.mode != 'EDIT_MESH'
-
-    def execute(self, context):
-        reaction_diffusion_add_handler(self, context)
-        set_animatable_fix_handler(self, context)
-
-        ob = context.object
-
-        ob.reaction_diffusion_settings.run = self.run
-        ob.reaction_diffusion_settings.dt = self.dt
-        ob.reaction_diffusion_settings.time_steps = self.time_steps
-        ob.reaction_diffusion_settings.f = self.f
-        ob.reaction_diffusion_settings.k = self.k
-        ob.reaction_diffusion_settings.diff_a = self.diff_a
-        ob.reaction_diffusion_settings.diff_b = self.diff_b
-
-        # check vertex group A
-        try:
-            vg = ob.vertex_groups['A']
-        except:
-            ob.vertex_groups.new(name='A')
-        # check vertex group B
-        try:
-            vg = ob.vertex_groups['B']
-        except:
-            ob.vertex_groups.new(name='B')
-
-        for v in ob.data.vertices:
-            ob.vertex_groups['A'].add([v.index], 1, 'REPLACE')
-            ob.vertex_groups['B'].add([v.index], 0, 'REPLACE')
-
-        ob.vertex_groups.update()
-        ob.data.update()
-        bpy.ops.object.mode_set(mode='WEIGHT_PAINT')
-
-        return {'FINISHED'}
-
-class reset_reaction_diffusion_weight(Operator):
-    bl_idname = "object.reset_reaction_diffusion_weight"
-    bl_label = "Reset Reaction Diffusion Weight"
-    bl_description = ("Set A and B weight to default values")
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context):
-        return context.object.type == 'MESH' and context.mode != 'EDIT_MESH'
-
-    def execute(self, context):
-        reaction_diffusion_add_handler(self, context)
-        set_animatable_fix_handler(self, context)
-
-        ob = context.object
-
-        # check vertex group A
-        try:
-            vg = ob.vertex_groups['A']
-        except:
-            ob.vertex_groups.new(name='A')
-        # check vertex group B
-        try:
-            vg = ob.vertex_groups['B']
-        except:
-            ob.vertex_groups.new(name='B')
-
-        for v in ob.data.vertices:
-            ob.vertex_groups['A'].add([v.index], 1, 'REPLACE')
-            ob.vertex_groups['B'].add([v.index], 0, 'REPLACE')
-
-        ob.vertex_groups.update()
-        ob.data.update()
-        bpy.ops.object.mode_set(mode='WEIGHT_PAINT')
-
-        return {'FINISHED'}
-
-class bake_reaction_diffusion(Operator):
-    bl_idname = "object.bake_reaction_diffusion"
-    bl_label = "Bake Data"
-    bl_description = ("Bake the Reaction-Diffusion to the cache directory")
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context):
-        return context.object.type == 'MESH' and context.mode != 'EDIT_MESH'
-
-    def execute(self, context):
-        ob = context.object
-        props = ob.reaction_diffusion_settings
-        if props.fast_bake:
-            bool_run = props.run
-            props.run = False
-            context.scene.frame_current = props.cache_frame_start
-            fast_bake_def(ob, frame_start=props.cache_frame_start, frame_end=props.cache_frame_end)
-            #create_fast_bake_def(ob, frame_start=props.cache_frame_start, frame_end=props.cache_frame_end)
-            context.scene.frame_current = props.cache_frame_end
-            props.run = bool_run
-        else:
-            for i in range(props.cache_frame_start, props.cache_frame_end):
-                context.scene.frame_current = i
-                reaction_diffusion_def(ob, bake=True)
-        props.bool_cache = True
-
-        return {'FINISHED'}
-
-class reaction_diffusion_free_data(Operator):
-    bl_idname = "object.reaction_diffusion_free_data"
-    bl_label = "Free Data"
-    bl_description = ("Free Reaction-Diffusion data")
-    bl_options = {'REGISTER', 'UNDO'}
-
-    @classmethod
-    def poll(cls, context):
-        return context.object.type == 'MESH'
-
-    def execute(self, context):
-        ob = context.object
-        props = ob.reaction_diffusion_settings
-        props.bool_cache = False
-
-        folder = Path(props.cache_dir)
-        for i in range(props.cache_frame_start, props.cache_frame_end):
-            data_a = folder / "a_{:04d}".format(i)
-            if os.path.exists(data_a):
-                os.remove(data_a)
-            data_a = folder / "b_{:04d}".format(i)
-            if os.path.exists(data_a):
-                os.remove(data_a)
-        return {'FINISHED'}
-
-from bpy.app.handlers import persistent
-
-def reaction_diffusion_scene(scene, bake=False):
-    for ob in scene.objects:
-        if ob.reaction_diffusion_settings.run:
-            reaction_diffusion_def(ob)
-
-def reaction_diffusion_def(ob, bake=False):
-
-    scene = bpy.context.scene
-    start = time.time()
-    if type(ob) == bpy.types.Scene: return None
-    props = ob.reaction_diffusion_settings
-
-    if bake or props.bool_cache or props.bake_geometry:
-        if props.cache_dir == '':
-            letters = string.ascii_letters
-            random_name = ''.join(rnd.choice(letters) for i in range(6))
-            if bpy.context.blend_data.filepath == '':
-                folder = Path(bpy.context.preferences.filepaths.temporary_directory)
-                folder = folder / 'reaction_diffusion_cache' / random_name
-            else:
-                folder = '//' + Path(bpy.context.blend_data.filepath).stem
-                folder = Path(bpy.path.abspath(folder)) / 'reaction_diffusion_cache' / random_name
-            folder.mkdir(parents=True, exist_ok=True)
-            props.cache_dir = str(folder)
-        else:
-            folder = Path(props.cache_dir)
-
-    me = ob.data
-    n_edges = len(me.edges)
-    n_verts = len(me.vertices)
-    a = np.zeros(n_verts)
-    b = np.zeros(n_verts)
-
-    print("{:6d} Reaction-Diffusion: {}".format(scene.frame_current, ob.name))
-
-    if not props.bool_cache:
-        if props.bool_mod:
-            # hide deforming modifiers
-            mod_visibility = []
-            for m in ob.modifiers:
-                mod_visibility.append(m.show_viewport)
-                if not mod_preserve_shape(m): m.show_viewport = False
-
-            # evaluated mesh
-            dg = bpy.context.evaluated_depsgraph_get()
-            ob_eval = ob.evaluated_get(dg)
-            me = bpy.data.meshes.new_from_object(ob_eval, preserve_all_data_layers=True, depsgraph=dg)
-
-            # set original visibility
-            for v, m in zip(mod_visibility, ob.modifiers):
-                m.show_viewport = v
-            ob.modifiers.update()
-
-        bm = bmesh.new()   # create an empty BMesh
-        bm.from_mesh(me)   # fill it in from a Mesh
-        dvert_lay = bm.verts.layers.deform.active
-
-        dt = props.dt
-        time_steps = props.time_steps
-        f = props.f
-        k = props.k
-        diff_a = props.diff_a
-        diff_b = props.diff_b
-        scale = props.diff_mult
-
-        brush_mult = props.brush_mult
-
-        # store weight values
-        if 'dB' in ob.vertex_groups: db = np.zeros(n_verts)
-        if 'grad' in ob.vertex_groups: grad = np.zeros(n_verts)
-
-        if props.vertex_group_diff_a != '': diff_a = np.zeros(n_verts)
-        if props.vertex_group_diff_b != '': diff_b = np.zeros(n_verts)
-        if props.vertex_group_scale != '': scale = np.zeros(n_verts)
-        if props.vertex_group_f != '': f = np.zeros(n_verts)
-        if props.vertex_group_k != '': k = np.zeros(n_verts)
-        if props.vertex_group_brush != '': brush = np.zeros(n_verts)
-        else: brush = 0
-
-        group_index_a = ob.vertex_groups["A"].index
-        group_index_b = ob.vertex_groups["B"].index
-        a = bmesh_get_weight_numpy(group_index_a, dvert_lay, bm.verts)
-        b = bmesh_get_weight_numpy(group_index_b, dvert_lay, bm.verts)
-
-        if props.vertex_group_diff_a != '':
-            group_index = ob.vertex_groups[props.vertex_group_diff_a].index
-            diff_a = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts, normalized=True)
-            if props.invert_vertex_group_diff_a:
-                vg_bounds = (props.min_diff_a, props.max_diff_a)
-            else:
-                vg_bounds = (props.max_diff_a, props.min_diff_a)
-            diff_a = np.interp(diff_a, (0,1), vg_bounds)
-
-        if props.vertex_group_diff_b != '':
-            group_index = ob.vertex_groups[props.vertex_group_diff_b].index
-            diff_b = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts, normalized=True)
-            if props.invert_vertex_group_diff_b:
-                vg_bounds = (props.max_diff_b, props.min_diff_b)
-            else:
-                vg_bounds = (props.min_diff_b, props.max_diff_b)
-            diff_b = np.interp(diff_b, (0,1), vg_bounds)
-
-        if props.vertex_group_scale != '':
-            group_index = ob.vertex_groups[props.vertex_group_scale].index
-            scale = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts, normalized=True)
-            if props.invert_vertex_group_scale:
-                vg_bounds = (props.max_scale, props.min_scale)
-            else:
-                vg_bounds = (props.min_scale, props.max_scale)
-            scale = np.interp(scale, (0,1), vg_bounds)
-
-        if props.vertex_group_f != '':
-            group_index = ob.vertex_groups[props.vertex_group_f].index
-            f = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts, normalized=True)
-            if props.invert_vertex_group_f:
-                vg_bounds = (props.max_f, props.min_f)
-            else:
-                vg_bounds = (props.min_f, props.max_f)
-            f = np.interp(f, (0,1), vg_bounds, )
-
-        if props.vertex_group_k != '':
-            group_index = ob.vertex_groups[props.vertex_group_k].index
-            k = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts, normalized=True)
-            if props.invert_vertex_group_k:
-                vg_bounds = (props.max_k, props.min_k)
-            else:
-                vg_bounds = (props.min_k, props.max_k)
-            k = np.interp(k, (0,1), vg_bounds)
-
-        if props.vertex_group_brush != '':
-            group_index = ob.vertex_groups[props.vertex_group_brush].index
-            brush = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts, normalized=True)
-            brush *= brush_mult
-
-        diff_a *= scale
-        diff_b *= scale
-
-        force_bake_geometry = False
-        if props.bake_geometry:
-            if props.update_baked_geometry:
-                force_bake_geometry = True
-            else:
-                try:
-                    file_name = folder / "edge_verts_0000"
-                    edge_verts = np.fromfile(file_name, dtype='int')
-                    file_name = folder / "field_mult_0000"
-                    field_mult = np.fromfile(file_name)
-                    n_edges = int(len(edge_verts)/2)
-                except:
-                    force_bake_geometry = True
-
-        if not props.bake_geometry or force_bake_geometry:
-            is_vector_field = True
-
-            if props.vector_field_mode not in ('NONE'):
-                if props.vector_field_mode == 'VECTOR':
-                    vec = Vector(props.vector)
-                    vector_field = [vec]*n_edges
-
-                if props.vector_field_mode == 'OBJECT':
-                    print(props.vector_field_object)
-                    if props.vector_field_object:
-                        mat = props.vector_field_object.matrix_world
-                    else:
-                        mat = ob.matrix_world
-                    vec = Vector((mat[0][2],mat[1][2],mat[2][2]))
-                    vector_field = [vec]*n_edges
-
-                if props.vector_field_mode == 'XYZ':
-                    vgk = ob.vertex_groups.keys()
-                    if 'x' in vgk and 'y' in vgk and 'z' in vgk:
-                        group_index = ob.vertex_groups["x"].index
-                        field_x = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-                        group_index = ob.vertex_groups["y"].index
-                        field_y = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-                        group_index = ob.vertex_groups["z"].index
-                        field_z = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-                        field_x = field_x*2-1
-                        field_y = field_y*2-1
-                        field_z = field_z*2-1
-                        vector_field = []
-                        for x,y,z in zip(field_x, field_y, field_z):
-                            vector_field.append(Vector((x,y,z)).normalized())
-                    else:
-                        is_vector_field = False
-
-                if props.vector_field_mode == 'GRADIENT':
-                    is_vector_field = False
-                    if props.vertex_group_gradient:
-                        if props.vertex_group_gradient in ob.vertex_groups.keys():
-                            edge_verts = [0]*n_edges*2
-                            me.edges.foreach_get("vertices", edge_verts)
-                            edge_verts = np.array(edge_verts)
-
-                            group_index = ob.vertex_groups[props.vertex_group_gradient].index
-                            weight = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-
-                            vector_field = []
-                            for v0 in bm.verts:
-                                vec = Vector((0,0,0))
-                                w0 = weight[v0.index]
-                                for e in v0.link_edges:
-                                    v1 = e.other_vert(v0)
-                                    dw = weight[v1.index]-w0
-                                    vec += (v1.co-v0.co)*dw
-                                vec.normalized()
-                                vector_field.append(vec)
-                            is_vector_field = True
-
-                if props.vector_field_mode == 'VECTOR_ATTRIBUTE':
-                    if 'VectorField' in ob.data.attributes:
-                        vectors_components = [0]*n_verts*3
-                        ob.data.attributes['VectorField'].data.foreach_get('vector', vectors_components)
-                        vector_field = []
-                        for i in range(n_verts):
-                            x = vectors_components[i*3]
-                            y = vectors_components[i*3+1]
-                            z = vectors_components[i*3+2]
-                            vector_field.append(Vector((x,y,z)).normalized())
-                    else:
-                        is_vector_field = False
-
-                if is_vector_field:
-                    edge_verts = []
-                    field_mult = []
-                    edge_verts_dict = {}
-                    for e in me.edges:
-                        pair = [e.vertices[0],e.vertices[1]]
-                        pair.sort()
-                        edge_verts_dict[str(pair[0]) + " " + str(pair[1])] = 0
-                    '''
-                    for p in me.polygons:
-                        n_face_verts = len(p.vertices)
-                        for i in range(n_face_verts-1):
-                            for j in range(i+1, n_face_verts):
-                                pair = [p.vertices[i],p.vertices[j]]
-                                pair.sort()
-                                edge_verts_dict[str(pair[0]) + " " + str(pair[1])] = 0
-                    '''
-
-                    if props.perp_vector_field:
-                        for i, vert in enumerate(bm.verts):
-                            vector_field[i] = vector_field[i].cross(vert.normal)
-                    for pair in edge_verts_dict.keys():
-                        pair = pair.split()
-                        id0 = int(pair[0])
-                        id1 = int(pair[1])
-                        edge_verts.append(id0)
-                        edge_verts.append(id1)
-                        v0 = me.vertices[id0].co
-                        v1 = me.vertices[id1].co
-                        vec = (v1-v0).normalized()
-                        mult0 = abs(vec.dot(vector_field[id0]))
-                        mult1 = abs(vec.dot(vector_field[id1]))
-                        field_mult.append((mult0 + mult1)/2)
-                    n_edges = len(edge_verts_dict.keys())-1
-                    field_mult = np.array(field_mult)
-            else:
-                is_vector_field = False
-
-            if props.vector_field_mode == 'NONE' or not is_vector_field:
-                edge_verts = [0]*n_edges*2
-                field_mult = np.zeros((1))
-                me.edges.foreach_get("vertices", edge_verts)
-
-        edge_verts = np.array(edge_verts)
-
-        if force_bake_geometry:
-            if not(os.path.exists(folder)):
-                os.mkdir(folder)
-            file_name = folder / "edge_verts_0000"
-            edge_verts.tofile(file_name)
-            file_name = folder / "field_mult_0000"
-            field_mult.tofile(file_name)
-            props.update_baked_geometry = False
-
-        field_mult = field_mult*props.anisotropy + (1-props.anisotropy)
-
-        timeElapsed = time.time() - start
-        print('       Preparation Time:',timeElapsed)
-        start = time.time()
-
-        try:
-            _f = f if type(f) is np.ndarray else np.array((f,))
-            _k = k if type(k) is np.ndarray else np.array((k,))
-            _diff_a = diff_a if type(diff_a) is np.ndarray else np.array((diff_a,))
-            _diff_b = diff_b if type(diff_b) is np.ndarray else np.array((diff_b,))
-            _brush = brush if type(brush) is np.ndarray else np.array((brush,))
-            if len(field_mult) == 1:
-                a, b = numba_reaction_diffusion(n_verts, n_edges, edge_verts, a, b, _brush, _diff_a, _diff_b, _f, _k, dt, time_steps)
-            else:
-                a, b = numba_reaction_diffusion_anisotropic(n_verts, n_edges, edge_verts, a, b, _brush, _diff_a, _diff_b, _f, _k, dt, time_steps, field_mult)
-        except:
-            print('Not using Numba! The simulation could be slow.')
-            arr = np.arange(n_edges)
-            id0 = edge_verts[arr*2]     # first vertex indices for each edge
-            id1 = edge_verts[arr*2+1]   # second vertex indices for each edge
-            if len(field_mult) == 1: mult = 1
-            else: mult = field_mult[arr]   # second vertex indices for each edge
-            for i in range(time_steps):
-                b += brush
-                lap_a = np.zeros(n_verts)
-                lap_b = np.zeros(n_verts)
-                lap_a0 =  (a[id1] -  a[id0])*mult   # laplacian increment for first vertex of each edge
-                lap_b0 =  (b[id1] -  b[id0])*mult   # laplacian increment for first vertex of each edge
-
-                np.add.at(lap_a, id0, lap_a0)
-                np.add.at(lap_b, id0, lap_b0)
-                np.add.at(lap_a, id1, -lap_a0)
-                np.add.at(lap_b, id1, -lap_b0)
-                ab2 = a*b**2
-                a += eval("(diff_a*lap_a - ab2 + f*(1-a))*dt")
-                b += eval("(diff_b*lap_b + ab2 - (k+f)*b)*dt")
-
-                a = nan_to_num(a)
-                b = nan_to_num(b)
-
-        timeElapsed = time.time() - start
-        print('       Simulation Time:',timeElapsed)
-
-    if bake:
-        if not(os.path.exists(folder)):
-            os.mkdir(folder)
-        file_name = folder / "a_{:04d}".format(scene.frame_current)
-        a.tofile(file_name)
-        file_name = folder / "b_{:04d}".format(scene.frame_current)
-        b.tofile(file_name)
-    elif props.bool_cache:
-        try:
-            file_name = folder / "a_{:04d}".format(scene.frame_current)
-            a = np.fromfile(file_name)
-            file_name = folder / "b_{:04d}".format(scene.frame_current)
-            b = np.fromfile(file_name)
-        except:
-            print('       Cannot read cache.')
-            return
-
-    if props.update_weight_a or props.update_weight_b:
-        start = time.time()
-        if props.update_weight_a:
-            if 'A' in ob.vertex_groups.keys():
-                vg_a = ob.vertex_groups['A']
-            else:
-                vg_a = ob.vertex_groups.new(name='A')
-        else:
-            vg_a = None
-        if props.update_weight_b:
-            if 'B' in ob.vertex_groups.keys():
-                vg_b = ob.vertex_groups['B']
-            else:
-                vg_b = ob.vertex_groups.new(name='B')
-        else:
-            vg_b = None
-        if vg_a == vg_b == None:
-            pass
-        else:
-            if ob.mode == 'WEIGHT_PAINT':# or props.bool_cache:
-                # slower, but prevent crashes
-                for i in range(n_verts):
-                    if vg_a: vg_a.add([i], a[i], 'REPLACE')
-                    if vg_b: vg_b.add([i], b[i], 'REPLACE')
-            else:
-                if props.bool_mod or props.bool_cache:
-                    #bm.free()               # release old bmesh
-                    bm = bmesh.new()        # create an empty BMesh
-                    bm.from_mesh(ob.data)   # fill it in from a Mesh
-                    dvert_lay = bm.verts.layers.deform.active
-                # faster, but can cause crashes while painting weight
-                if vg_a: index_a = vg_a.index
-                if vg_b: index_b = vg_b.index
-                for i, v in enumerate(bm.verts):
-                    dvert = v[dvert_lay]
-                    if vg_a: dvert[index_a] = a[i]
-                    if vg_b: dvert[index_b] = b[i]
-                bm.to_mesh(ob.data)
-                bm.free()
-        print('       Writing Vertex Groups Time:',time.time() - start)
-    if props.normalize:
-        min_a = np.min(a)
-        max_a = np.max(a)
-        min_b = np.min(b)
-        max_b = np.max(b)
-        a = (a - min_a)/(max_a - min_a)
-        b = (b - min_b)/(max_b - min_b)
-    split_a = None
-    split_b = None
-    splitted = False
-    if props.update_colors:#_a or props.update_colors_b:
-        start = time.time()
-        loops_size = get_attribute_numpy(me.polygons, attribute='loop_total', mult=1)
-        n_colors = np.sum(loops_size)
-        v_id = np.ones(n_colors)
-        me.polygons.foreach_get('vertices',v_id)
-        v_id = v_id.astype(int)
-        split_a = a[v_id,None]
-        split_b = b[v_id,None]
-        splitted = True
-        ones = np.ones((n_colors,1))
-        rgba = np.concatenate((split_a,split_b,ones,ones),axis=1).flatten()
-        if 'AB' in ob.data.vertex_colors.keys():
-            vc = ob.data.vertex_colors['AB']
-        else:
-            vc = ob.data.vertex_colors.new(name='AB')
-        vc.data.foreach_set('color',rgba)
-        ob.data.vertex_colors.update()
-
-        print('       Writing Vertex Colors Time:',time.time() - start)
-    if props.update_uv:
-        start = time.time()
-        if 'AB' in me.uv_layers.keys():
-            uv_layer = me.uv_layers['AB']
-        else:
-            uv_layer = me.uv_layers.new(name='AB')
-        if not splitted:
-            loops_size = get_attribute_numpy(me.polygons, attribute='loop_total', mult=1)
-            n_data = np.sum(loops_size)
-            v_id = np.ones(n_data)
-            me.polygons.foreach_get('vertices',v_id)
-            v_id = v_id.astype(int)
-            split_a = a[v_id,None]
-            split_b = b[v_id,None]
-        uv = np.concatenate((split_a,split_b),axis=1).flatten()
-        uv_layer.data.foreach_set('uv',uv)
-        me.uv_layers.update()
-        print('       Writing UV Map Time:',time.time() - start)
-
-    for ps in ob.particle_systems:
-        if ps.vertex_group_density == 'B' or ps.vertex_group_density == 'A':
-            ps.invert_vertex_group_density = not ps.invert_vertex_group_density
-            ps.invert_vertex_group_density = not ps.invert_vertex_group_density
-
-    if props.bool_mod and not props.bool_cache: bpy.data.meshes.remove(me)
-
-def fast_bake_def(ob, frame_start=1, frame_end=250):
-    scene = bpy.context.scene
-    start = time.time()
-    if type(ob) == bpy.types.Scene: return None
-    props = ob.reaction_diffusion_settings
-
-    # Define cache folder
-    if props.cache_dir == '':
-        letters = string.ascii_letters
-        random_name = ''.join(rnd.choice(letters) for i in range(6))
-        if bpy.context.blend_data.filepath == '':
-            folder = Path(bpy.context.preferences.filepaths.temporary_directory)
-            folder = folder / 'reaction_diffusion_cache' / random_name
-        else:
-            folder = '//' + Path(bpy.context.blend_data.filepath).stem
-            folder = Path(bpy.path.abspath(folder)) / 'reaction_diffusion_cache' / random_name
-        folder.mkdir(parents=True, exist_ok=True)
-        props.cache_dir = str(folder)
-    else:
-        folder = Path(props.cache_dir)
-
-    if props.bool_mod:
-        # hide deforming modifiers
-        mod_visibility = []
-        for m in ob.modifiers:
-            mod_visibility.append(m.show_viewport)
-            if not mod_preserve_shape(m): m.show_viewport = False
-
-        # evaluated mesh
-        dg = bpy.context.evaluated_depsgraph_get()
-        ob_eval = ob.evaluated_get(dg)
-        me = bpy.data.meshes.new_from_object(ob_eval, preserve_all_data_layers=True, depsgraph=dg)
-
-        # set original visibility
-        for v, m in zip(mod_visibility, ob.modifiers):
-            m.show_viewport = v
-        ob.modifiers.update()
-    else:
-        me = ob.data
-
-    bm = bmesh.new()   # create an empty BMesh
-    bm.from_mesh(me)   # fill it in from a Mesh
-    dvert_lay = bm.verts.layers.deform.active
-    n_edges = len(me.edges)
-    n_verts = len(me.vertices)
-    a = np.zeros(n_verts)
-    b = np.zeros(n_verts)
-    group_index_a = ob.vertex_groups["A"].index
-    group_index_b = ob.vertex_groups["B"].index
-
-    dt = props.dt
-    time_steps = props.time_steps
-    f = props.f
-    k = props.k
-    diff_a = props.diff_a
-    diff_b = props.diff_b
-    scale = props.diff_mult
-
-    brush_mult = props.brush_mult
-
-    # store weight values
-    if 'dB' in ob.vertex_groups: db = np.zeros(n_verts)
-    if 'grad' in ob.vertex_groups: grad = np.zeros(n_verts)
-
-    if props.vertex_group_diff_a != '': diff_a = np.zeros(n_verts)
-    if props.vertex_group_diff_b != '': diff_b = np.zeros(n_verts)
-    if props.vertex_group_scale != '': scale = np.zeros(n_verts)
-    if props.vertex_group_f != '': f = np.zeros(n_verts)
-    if props.vertex_group_k != '': k = np.zeros(n_verts)
-    if props.vertex_group_brush != '': brush = np.zeros(n_verts)
-    else: brush = 0
-
-    a = bmesh_get_weight_numpy(group_index_a, dvert_lay, bm.verts)
-    b = bmesh_get_weight_numpy(group_index_b, dvert_lay, bm.verts)
-
-    if props.vertex_group_diff_a != '':
-        group_index = ob.vertex_groups[props.vertex_group_diff_a].index
-        diff_a = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-        if props.invert_vertex_group_diff_a:
-            vg_bounds = (props.min_diff_a, props.max_diff_a)
-        else:
-            vg_bounds = (props.max_diff_a, props.min_diff_a)
-        diff_a = np.interp(diff_a, (0,1), vg_bounds)
-
-    if props.vertex_group_diff_b != '':
-        group_index = ob.vertex_groups[props.vertex_group_diff_b].index
-        diff_b = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-        if props.invert_vertex_group_diff_b:
-            vg_bounds = (props.max_diff_b, props.min_diff_b)
-        else:
-            vg_bounds = (props.min_diff_b, props.max_diff_b)
-        diff_b = np.interp(diff_b, (0,1), vg_bounds)
-
-    if props.vertex_group_scale != '':
-        group_index = ob.vertex_groups[props.vertex_group_scale].index
-        scale = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-        if props.invert_vertex_group_scale:
-            vg_bounds = (props.max_scale, props.min_scale)
-        else:
-            vg_bounds = (props.min_scale, props.max_scale)
-        scale = np.interp(scale, (0,1), vg_bounds)
-
-    if props.vertex_group_f != '':
-        group_index = ob.vertex_groups[props.vertex_group_f].index
-        f = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-        if props.invert_vertex_group_f:
-            vg_bounds = (props.max_f, props.min_f)
-        else:
-            vg_bounds = (props.min_f, props.max_f)
-        f = np.interp(f, (0,1), vg_bounds, )
-
-    if props.vertex_group_k != '':
-        group_index = ob.vertex_groups[props.vertex_group_k].index
-        k = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-        if props.invert_vertex_group_k:
-            vg_bounds = (props.max_k, props.min_k)
-        else:
-            vg_bounds = (props.min_k, props.max_k)
-        k = np.interp(k, (0,1), vg_bounds)
-
-    if props.vertex_group_brush != '':
-        group_index = ob.vertex_groups[props.vertex_group_brush].index
-        brush = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-        brush *= brush_mult
-
-    diff_a *= scale
-    diff_b *= scale
-
-    if True:
-
-        group_index = ob.vertex_groups["x"].index
-        field_x = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-        group_index = ob.vertex_groups["y"].index
-        field_y = bmesh_get_weight_numpy(group_index, dvert_lay, bm.verts)
-        field_x = field_x*2-1
-        field_y = field_y*2-1
-        vector_field = []
-        for x,y in zip(field_x, field_y):
-            vector_field.append(Vector((x,y,0)))
-        edge_verts = []
-        field_mult = []
-        edge_verts_dict = {}
-        for p in me.polygons:
-            n_face_verts = len(p.vertices)
-            for i in range(n_face_verts-1):
-                for j in range(i+1, n_face_verts):
-                    pair = [p.vertices[i],p.vertices[j]]
-                    pair.sort()
-                    edge_verts_dict[str(pair[0]) + " " + str(pair[1])] = 0
-        for pair in edge_verts_dict.keys():
-            pair = pair.split()
-            v0 = int(pair[0])
-            v1 = int(pair[1])
-            field_dir = (vector_field[v0]+vector_field[v1]).normalized()
-            edge_verts.append(v0)
-            edge_verts.append(v1)
-            v0 = me.vertices[v0].co
-            v1 = me.vertices[v1].co
-            vec = (v1-v0).normalized()
-            #field_dir = Vector((0,0,1))
-            field_mult.append(abs(vec.dot(field_dir))*0.5 + 0.5)
-        n_edges = len(edge_verts_dict.keys())-1
-        field_mult = np.array(field_mult)
-    else:
-        edge_verts = [0]*n_edges*2
-        me.edges.foreach_get("vertices", edge_verts)
-    edge_verts = np.array(edge_verts)
-
-    timeElapsed = time.time() - start
-    print('       Preparation Time:',timeElapsed)
-    start = time.time()
-
-    try:
-        edge_verts = np.array(edge_verts)
-        _f = f if type(f) is np.ndarray else np.array((f,))
-        _k = k if type(k) is np.ndarray else np.array((k,))
-        _diff_a = diff_a if type(diff_a) is np.ndarray else np.array((diff_a,))
-        _diff_b = diff_b if type(diff_b) is np.ndarray else np.array((diff_b,))
-        _brush = brush if type(brush) is np.ndarray else np.array((brush,))
-
-        run_rd = False
-        for j in range(props.cache_frame_start, props.cache_frame_end+1):
-            start2 = time.time()
-            print("{:6d} Reaction-Diffusion: {}".format(j, ob.name))
-            if run_rd:
-                b += _brush
-                a, b = numba_reaction_diffusion(n_verts, n_edges, edge_verts, a, b, _brush, _diff_a, _diff_b, _f, _k, dt, time_steps, field_mult)
-            else:
-                run_rd = True
-
-            if not(os.path.exists(folder)):
-                os.mkdir(folder)
-            file_name = folder / "a_{:04d}".format(j)
-            a.tofile(file_name)
-            file_name = folder / "b_{:04d}".format(j)
-            b.tofile(file_name)
-
-            timeElapsed = time.time() - start2
-            print('       Simulation Time:',timeElapsed)
-
-    except:
-        print('Not using Numba! The simulation could be slow.')
-        edge_verts = np.array(edge_verts)
-        arr = np.arange(n_edges)*2
-        id0 = edge_verts[arr]     # first vertex indices for each edge
-        id1 = edge_verts[arr+1]   # second vertex indices for each edge
-        for j in range(props.cache_frame_start, props.cache_frame_end):
-            for i in range(time_steps):
-                b += brush
-                lap_a = np.zeros(n_verts)
-                lap_b = np.zeros(n_verts)
-                lap_a0 =  a[id1] -  a[id0]   # laplacian increment for first vertex of each edge
-                lap_b0 =  b[id1] -  b[id0]   # laplacian increment for first vertex of each edge
-
-                np.add.at(lap_a, id0, lap_a0)
-                np.add.at(lap_b, id0, lap_b0)
-                np.add.at(lap_a, id1, -lap_a0)
-                np.add.at(lap_b, id1, -lap_b0)
-
-                ab2 = a*b**2
-                a += eval("(diff_a*lap_a - ab2 + f*(1-a))*dt")
-                b += eval("(diff_b*lap_b + ab2 - (k+f)*b)*dt")
-
-                a = nan_to_num(a)
-                b = nan_to_num(b)
-
-            if not(os.path.exists(folder)):
-                os.mkdir(folder)
-            file_name = folder / "a_{:04d}".format(j)
-            a.tofile(file_name)
-            file_name = folder / "b_{:04d}".format(j)
-            b.tofile(file_name)
-
-    if ob.mode == 'WEIGHT_PAINT':
-        # slower, but prevent crashes
-        vg_a = ob.vertex_groups['A']
-        vg_b = ob.vertex_groups['B']
-        for i in range(n_verts):
-            vg_a.add([i], a[i], 'REPLACE')
-            vg_b.add([i], b[i], 'REPLACE')
-    else:
-        if props.bool_mod:
-            bm.free()               # release old bmesh
-            bm = bmesh.new()        # create an empty BMesh
-            bm.from_mesh(ob.data)   # fill it in from a Mesh
-            dvert_lay = bm.verts.layers.deform.active
-        # faster, but can cause crashes while painting weight
-        for i, v in enumerate(bm.verts):
-            dvert = v[dvert_lay]
-            dvert[group_index_a] = a[i]
-            dvert[group_index_b] = b[i]
-        bm.to_mesh(ob.data)
-
-    # Update Vertex Colors
-    if 'A' in ob.data.vertex_colors or 'B' in ob.data.vertex_colors:
-        v_id = np.array([v for p in ob.data.polygons for v in p.vertices])
-
-        if 'B' in ob.data.vertex_colors:
-            c_val = b[v_id]
-            c_val = np.repeat(c_val, 4, axis=0)
-            vc = ob.data.vertex_colors['B']
-            vc.data.foreach_set('color',c_val.tolist())
-
-        if 'A' in ob.data.vertex_colors:
-            c_val = a[v_id]
-            c_val = np.repeat(c_val, 4, axis=0)
-            vc = ob.data.vertex_colors['A']
-            vc.data.foreach_set('color',c_val.tolist())
-
-    for ps in ob.particle_systems:
-        if ps.vertex_group_density == 'B' or ps.vertex_group_density == 'A':
-            ps.invert_vertex_group_density = not ps.invert_vertex_group_density
-            ps.invert_vertex_group_density = not ps.invert_vertex_group_density
-
-    if props.bool_mod: bpy.data.meshes.remove(me)
-    bm.free()
-    timeElapsed = time.time() - start
-    print('       Closing Time:',timeElapsed)
-
-def create_fast_bake_def(ob, frame_start=1, frame_end=250):
-    scene = bpy.context.scene
-    start = time.time()
-    if type(ob) == bpy.types.Scene: return None
-    props = ob.reaction_diffusion_settings
-
-    dt = props.dt
-    time_steps = props.time_steps
-    scale = props.diff_mult
-
-    if props.cache_dir == '':
-        letters = string.ascii_letters
-        random_name = ''.join(rnd.choice(letters) for i in range(6))
-        if bpy.context.blend_data.filepath == '':
-            folder = Path(bpy.context.preferences.filepaths.temporary_directory)
-            folder = folder / 'reaction_diffusion_cache' / random_name
-        else:
-            folder = '//' + Path(bpy.context.blend_data.filepath).stem
-            folder = Path(bpy.path.abspath(folder)) / 'reaction_diffusion_cache' / random_name
-        folder.mkdir(parents=True, exist_ok=True)
-        props.cache_dir = str(folder)
-    else:
-        folder = Path(props.cache_dir)
-
-    if props.bool_mod:
-        # hide deforming modifiers
-        mod_visibility = []
-        for m in ob.modifiers:
-            mod_visibility.append(m.show_viewport)
-            if not mod_preserve_shape(m): m.show_viewport = False
-
-        # evaluated mesh
-        dg = bpy.context.evaluated_depsgraph_get()
-        ob_eval = ob.evaluated_get(dg)
-        me = bpy.data.meshes.new_from_object(ob_eval, preserve_all_data_layers=True, depsgraph=dg)
-
-        # set original visibility
-        for v, m in zip(mod_visibility, ob.modifiers):
-            m.show_viewport = v
-        ob.modifiers.update()
-    else:
-        me = ob.data
-
-    bm = bmesh.new()   # create an empty BMesh
-    bm.from_mesh(me)   # fill it in from a Mesh
-    verts = get_vertices_numpy(me)
-    dvert_lay = bm.verts.layers.deform.active
-    n_edges = len(me.edges)
-    n_verts = len(me.vertices)
-    group_index_x = ob.vertex_groups["x"].index
-    group_index_y = ob.vertex_groups["y"].index
-    group_index_module = ob.vertex_groups["module"].index
-    group_index_values = ob.vertex_groups["values"].index
-
-    if not props.bool_cache:
-        time_steps = props.time_steps
-
-        # store weight values
-        if 'dB' in ob.vertex_groups: db = np.zeros(n_verts)
-        if 'grad' in ob.vertex_groups: grad = np.zeros(n_verts)
-        vec_x = np.zeros(n_verts)
-        vec_y = np.zeros(n_verts)
-        vec_module = np.zeros(n_verts)
-        values = np.zeros(n_verts)
-
-        vec_x = bmesh_get_weight_numpy(group_index_x, dvert_lay, bm.verts)
-        vec_y = bmesh_get_weight_numpy(group_index_y, dvert_lay, bm.verts)
-        vec_module = bmesh_get_weight_numpy(group_index_module, dvert_lay, bm.verts)
-        values = bmesh_get_weight_numpy(group_index_values, dvert_lay, bm.verts)
-        field = np.concatenate((vec_x[:,None],vec_y[:,None],vec_y[:,None]*0),axis=1)
-        field = field*2-1
-        field[:,2] = 0
-        edge_verts = get_edges_numpy(me)
-
-        id0 = edge_verts[:,0]
-        id1 = edge_verts[:,1]
-        vert0 = verts[id0]
-        vert1 = verts[id1]
-        vec = vert1-vert0
-        edge_field = (field[id0] + field[id1])/2    # average vector associated to the edge
-        # normalize vectors
-        vec /= np.linalg.norm(vec,axis=1)[:,None]
-        edge_field /= np.linalg.norm(edge_field,axis=1)[:,None]
-        edge_flow = np.einsum('...j,...j', vec, edge_field)
-
-        timeElapsed = time.time() - start
-        print('       Preparation Time:',timeElapsed)
-        start = time.time()
-
-        # Preserve energy
-        mult = np.zeros(values.shape)
-        np.add.at(mult,id0,-edge_flow)
-        np.add.at(mult,id1,edge_flow)
-        mult = scale/mult
-
-        bool_run = False
-        for j in range(props.cache_frame_start, props.cache_frame_end+1):
-            start2 = time.time()
-            print("{:6d} Reaction-Diffusion: {}".format(j, ob.name))
-            if bool_run:
-                values = integrate_field(n_edges,id0,id1,values,edge_flow,mult,time_steps)
-            else:
-                bool_run = True
-
-            if not(os.path.exists(folder)):
-                os.mkdir(folder)
-            file_name = folder / "a_{:04d}".format(j)
-            values.tofile(file_name)
-            file_name = folder / "b_{:04d}".format(j)
-            values.tofile(file_name)
-
-
-            timeElapsed = time.time() - start2
-            print('       Simulation Time:',timeElapsed)
-
-    if props.bool_mod: bpy.data.meshes.remove(me)
-    bm.free()
-    timeElapsed = time.time() - start
-    print('       Closing Time:',timeElapsed)
-
-
-
-
-
-class TISSUE_PT_reaction_diffusion(Panel):
-    bl_space_type = 'PROPERTIES'
-    bl_region_type = 'WINDOW'
-    bl_context = "data"
-    bl_label = "Tissue Reaction-Diffusion"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    @classmethod
-    def poll(cls, context):
-        return 'A' and 'B' in context.object.vertex_groups
-
-    def draw(self, context):
-        reaction_diffusion_add_handler(self, context)
-
-        ob = context.object
-        props = ob.reaction_diffusion_settings
-        layout = self.layout
-        col = layout.column(align=True)
-        row = col.row(align=True)
-        if not ("A" and "B" in ob.vertex_groups):
-            row.operator("object.start_reaction_diffusion",
-                        icon="EXPERIMENTAL",
-                        text="Reaction-Diffusion")
-        else:
-            row.operator("object.start_reaction_diffusion",
-                        icon="EXPERIMENTAL",
-                        text="Reset Reaction-Diffusion")
-            row = col.row(align=True)
-            row.prop(props, "run", text="Run Reaction-Diffusion")
-            col = layout.column(align=True)
-            row = col.row(align=True)
-            row.prop(props, "time_steps")
-            row.prop(props, "dt")
-            row.enabled = not props.bool_cache
-            col.separator()
-            row = col.row(align=True)
-            col1 = row.column(align=True)
-            col1.prop(props, "diff_a")
-            col1.enabled = props.vertex_group_diff_a == '' and not props.bool_cache
-            col1 = row.column(align=True)
-            col1.prop(props, "diff_b")
-            col1.enabled = props.vertex_group_diff_b == '' and not props.bool_cache
-            row = col.row(align=True)
-            row.prop(props, "diff_mult")
-            row.enabled = props.vertex_group_scale == '' and not props.bool_cache
-            #col.separator()
-            row = col.row(align=True)
-            col1 = row.column(align=True)
-            col1.prop(props, "f")
-            col1.enabled = props.vertex_group_f == '' and not props.bool_cache
-            col1 = row.column(align=True)
-            col1.prop(props, "k")
-            col1.enabled = props.vertex_group_k == '' and not props.bool_cache
-            col.separator()
-            col.label(text='Vector Field:')
-            col.prop(props, "vector_field_mode", text="Mode")
-            if props.vector_field_mode == 'OBJECT':
-                col.prop_search(props, "vector_field_object", context.scene, "objects", text='Object')
-            if props.vector_field_mode == 'GRADIENT':
-                col.prop_search(props, 'vertex_group_gradient', ob, "vertex_groups")
-            if props.vector_field_mode == 'XYZ':
-                vgk = ob.vertex_groups.keys()
-                if 'x' not in vgk:
-                    col.label(text="Vertex Group 'x' is missing", icon='ERROR')
-                if 'y' not in vgk:
-                    col.label(text="Vertex Group 'y' is missing", icon='ERROR')
-                if 'z' not in vgk:
-                    col.label(text="Vertex Group 'z' is missing", icon='ERROR')
-            if props.vector_field_mode == 'VECTOR_ATTRIBUTE':
-                vgk = ob.vertex_groups.keys()
-                if 'VectorField' not in ob.data.attributes:
-                    col.label(text="Vector Attribute 'VectorField' is missing", icon='ERROR')
-            if props.vector_field_mode == 'VECTOR':
-                row = col.row()
-                row.prop(props, "vector")
-            if props.vector_field_mode != 'NONE':
-                col.separator()
-                row = col.row()
-                row.prop(props, 'perp_vector_field', text='Perpendicular')
-                row.prop(props, "anisotropy")
-            col.separator()
-            col.label(text='Cache:')
-            col.prop(props, "cache_dir", text='')
-            col.separator()
-            row = col.row()
-            row.prop(props, "bake_geometry", icon="MESH_DATA")
-            file = bpy.context.blend_data.filepath
-            temp = bpy.context.preferences.filepaths.temporary_directory
-            if file == temp == props.cache_dir == '':
-                row.enabled = False
-            col.separator()
-            row = col.row(align=True)
-            row.prop(props, "cache_frame_start")
-            row.prop(props, "cache_frame_end")
-            col.separator()
-            if props.bool_cache:
-                col.operator("object.reaction_diffusion_free_data")
-            else:
-                row = col.row(align=True)
-                row.operator("object.bake_reaction_diffusion")
-                file = bpy.context.blend_data.filepath
-                temp = bpy.context.preferences.filepaths.temporary_directory
-                if file == temp == props.cache_dir == '':
-                    row.enabled = False
-                    col.label(text="Cannot use cache", icon='ERROR')
-                    col.label(text='please save the Blender or set a Cache directory')
-                col.prop(props, "fast_bake")
-
-            col.separator()
-            col.label(text='Output attributes:')
-            row = col.row(align=True)
-            col2 = row.column(align=True)
-            row2 = col2.row(align=True)
-            row2.prop(props, "update_weight_a", icon='GROUP_VERTEX', text='A')
-            row2.prop(props, "update_weight_b", icon='GROUP_VERTEX', text='B')
-            col2.enabled = props.bool_cache
-            row.separator()
-            row.prop(props, "update_colors", icon='GROUP_VCOL', text='AB')
-            row.separator()
-            row.prop(props, "update_uv", icon='GROUP_UVS', text='AB')
-            col.prop(props,'normalize')
-
-class TISSUE_PT_reaction_diffusion_weight(Panel):
-    bl_space_type = 'PROPERTIES'
-    bl_region_type = 'WINDOW'
-    bl_context = "data"
-    bl_parent_id = "TISSUE_PT_reaction_diffusion"
-    bl_label = "Vertex Groups"
-    bl_options = {'DEFAULT_CLOSED'}
-
-    @classmethod
-    def poll(cls, context):
-        return 'A' and 'B' in context.object.vertex_groups
-
-    def draw(self, context):
-        ob = context.object
-        props = ob.reaction_diffusion_settings
-        layout = self.layout
-        col = layout.column(align=True)
-        col.prop(props, "bool_mod")
-        if props.bool_mod and props.fast_bake:
-            col.label(text="When Fast Bake is on, the modifiers", icon='ERROR')
-            col.label(text="          are used only for the first frame")
-        col.separator()
-        insert_weight_parameter(col, ob, 'brush', text='Brush:')
-        insert_weight_parameter(col, ob, 'diff_a', text='Diff A:')
-        insert_weight_parameter(col, ob, 'diff_b', text='Diff B:')
-        insert_weight_parameter(col, ob, 'scale', text='Scale:')
-        insert_weight_parameter(col, ob, 'f', text='f:')
-        insert_weight_parameter(col, ob, 'k', text='k:')
-        col.enabled = not props.bool_cache
-
-def insert_weight_parameter(col, ob, name, text=''):
-    props = ob.reaction_diffusion_settings
-    split = col.split(factor=0.25, align=True)
-    col2 = split.column(align=True)
-    col2.label(text=text)
-    col2 = split.column(align=True)
-    row2 = col2.row(align=True)
-    row2.prop_search(props, 'vertex_group_' + name, ob, "vertex_groups", text='')
-    if name != 'brush':
-        row2.prop(props, "invert_vertex_group_" + name, text="", toggle=True, icon='ARROW_LEFTRIGHT')
-    if 'vertex_group_' + name in props:
-        if props['vertex_group_' + name] != '':
-            if name == 'brush':
-                col2.prop(props, "brush_mult")
-            else:
-                row2 = col2.row(align=True)
-                row2.prop(props, "min_" + name, text="Min")
-                row2 = col2.row(align=True)
-                row2.prop(props, "max_" + name, text="Max")
-    col.separator()
-
 
 def contour_bmesh(me, bm, weight, iso_val):
     bm.verts.ensure_lookup_table()
